requirements.txt
flask>=3.0.0
flask-sqlalchemy>=3.1.0
flask-migrate>=4.0.0
python-dotenv>=1.0.0
requests>=2.28.0
